+++ app/src/test/java/org/mozilla/fenix/perf/StrictModeManagerTest.kt
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.fenix
+package org.mozilla.fenix.perf
 
 import android.os.StrictMode
 import androidx.fragment.app.FragmentManager
@@ -21,6 +21,8 @@ import org.junit.Assert.fail
 import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
+import org.mozilla.fenix.Config
+import org.mozilla.fenix.ReleaseChannel
 import org.mozilla.fenix.components.Components
 import org.mozilla.fenix.helpers.FenixRobolectricTestRunner
